DDS2-592 - Update scaffolds
