+++ .g8/checkboxPage/generated-test/forms/$section$/$className$FormProviderSpec.scala
