+++ .g8/checkboxPage/generated-test/models/$section$/$className$Spec.scala
