+++ .g8/checkboxPage/generated-test/pages/$section$/$className$PageSpec.scala
