+++ .g8/intPage/generated-test/pages/$section$/$className$PageSpec.scala
