+++ .g8/stringPage/generated-test/pages/$section$/$className$PageSpec.scala
