+++ .g8/yesNoPage/generated-test/pages/$section$/$className$PageSpec.scala
