+++ test-utils/generators/PageGenerators.scala
@@ -20,10 +20,11 @@ import org.scalacheck.Arbitrary
 import pages._
 
 trait PageGenerators {
-
+  
   implicit lazy val arbitraryrelatesToPage: Arbitrary[RelatesToPage.type] =
     Arbitrary(RelatesToPage)
  
   implicit lazy val arbitraryReceivedALetterPage: Arbitrary[ReceivedALetterPage.type] =
     Arbitrary(ReceivedALetterPage)
+    
 }
